requirements.txt
Django==1.10
psycopg2==2.6.2
SQLAlchemy==1.0.14
django-bootstrap-pagination==1.6.2
djangorestframework==3.4.6
markdown==2.6.6
django-filter==0.14.0
